+++ lib/rubygems_pq_tls_policy/patch.rb
@@ -2,43 +2,35 @@
 
 module Gem
   module PqTlsPolicy
-    module Patch
-      def post_connection_check(hostname)
-        result = super
-
-        config = Gem::PqTlsPolicy.config
-        return result unless config.enabled?
-
-        config.validate!
-
-        group = respond_to?(:group) ? self.group : nil
-
-        if config.trace?
-          warn "[rubygems:tls] host=#{hostname} version=#{safe_ssl_version} cipher=#{safe_cipher_name} group=#{group.inspect}"
-        end
-
-        if config.pq_required? && !config.allowed_groups.include?(group)
-          raise Gem::PqTlsPolicy::Violation,
-            "RubyGems gem-server TLS key exchange policy violation: " \
-            "host=#{hostname.inspect} used group=#{group.inspect}; " \
-            "allowed=#{config.allowed_groups.join(',')}"
-        end
-
-        result
+    module PerConnectionCheck
+      def connect
+        super
+        Gem::PqTlsPolicy.check_net_http_connection!(self)
+      rescue Gem::PqTlsPolicy::Error
+        close_pq_tls_policy_socket
+        raise
       end
 
+      private :connect
+
+      def close_pq_tls_policy_socket
+        socket = instance_variable_get(:@socket)
+        socket.close if socket.respond_to?(:close)
+        instance_variable_set(:@socket, nil)
+      rescue StandardError
+        nil
+      end
+
+      private :close_pq_tls_policy_socket
+    end
+
+    module RequestHTTPSPoolPatch
       private
 
-      def safe_ssl_version
-        ssl_version
-      rescue StandardError
-        "unknown"
-      end
-
-      def safe_cipher_name
-        cipher&.first
-      rescue StandardError
-        "unknown"
+      def setup_connection(connection)
+        connection = Gem::Request.configure_connection_for_https(connection, @cert_files)
+        Gem::PqTlsPolicy.extend_net_http_connection!(connection)
+        Gem::Request::HTTPPool.instance_method(:setup_connection).bind_call(self, connection)
       end
     end
   end